cnn_model/requirements.txt
# ResNet-CBAM SAR Landcover Classification Requirements

# Core PyTorch and ML libraries
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
scikit-learn>=0.24.0

# Data handling and visualization
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.2.0
pillow>=8.0.0

# Geospatial data handling
rasterio>=1.2.0
gdal>=3.0.0

# Progress bars and utilities
tqdm>=4.60.0
opencv-python>=4.5.0

# Logging and monitoring
tensorboard>=2.4.0

# Configuration and serialization
dataclasses-json>=0.5.0

# Optional: For enhanced functionality
# jupyter>=1.0.0  # For notebook examples
# plotly>=5.0.0   # For interactive visualizations
# wandb>=0.12.0   # For experiment tracking